requirements.txt
# --- LiveKit Stack (2025 Stable) ---
livekit-agents>=1.3.0
livekit-plugins-silero>=0.7.0
livekit-plugins-openai>=0.10.0

# --- STT & TTS ---
faster-whisper>=1.0.3
parler-tts @ git+https://github.com/huggingface/parler-tts.git
transformers>=4.44.0
torch>=2.2.0
torchaudio>=2.2.0
soundfile>=0.12.1
tokenizers>=0.20.2

# --- LLM & Orchestration ---
ollama>=0.3.0

# --- Backend & Infrastructure ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
aiosqlite>=0.20.0
loguru>=0.7.2
python-dotenv>=1.0.1
httpx>=0.27.0
numpy>=1.26.4